+++ src/zenml/actions/pipeline_run/pipeline_run_action.py
@@ -33,8 +33,8 @@ from zenml.models import (
 )
 from zenml.models.v2.base.base import BaseResponse
 from zenml.zen_server.auth import AuthContext
-from zenml.zen_server.pipeline_deployment.utils import run_template
 from zenml.zen_server.rbac.models import ResourceType
+from zenml.zen_server.template_execution.utils import run_template
 from zenml.zen_server.utils import server_config
 
 logger = get_logger(__name__)
